+++ src/Saritasa.Tools.Messages/Common/Repositories/QueryProviders/PostgreSqlQueryProvider.cs
@@ -118,7 +118,7 @@ namespace Saritasa.Tools.Messages.Common.Repositories.QueryProviders
             }
             if (messageQuery.Status != null)
             {
-                ssb.Where("status").EqualsTo(messageQuery.Status);
+                ssb.Where("status").EqualsTo((byte)messageQuery.Status);
             }
             if (messageQuery.Type != null)
             {
